pyproject.toml
[project]
name = "barca-calendar-bot"
version = "3.0.0"
description = "Bot que sincroniza el calendario del FC Barcelona con Google Calendar, añadiendo probabilidades de victoria y resúmenes automáticos."
readme = "README.md"
requires-python = ">=3.12"
license = {text = "MIT"}
authors = [
    {name = "Aleix Quirante", email = "quirante70@gmail.com"},
]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "Topic :: Utilities",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3.12",
]

dependencies = [
    "requests>=2.32",
    "icalendar>=5.0",
    "google-api-python-client>=2.130",
    "google-auth-httplib2>=0.2",
    "google-auth-oauthlib>=1.2",
    "beautifulsoup4>=4.12",
    "lxml>=5.0",
    "pydantic>=2.10",
    "pydantic-settings>=2.5",
    "httpx>=0.27",
    "cachetools>=5.3",
    "feedparser>=6.0",
    "openai>=1.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-cov>=5.0",
    "black>=24.0",
    "ruff>=0.4",
    "mypy>=1.10",
    "types-requests",
    "types-cachetools",
]

[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[tool.black]
line-length = 88
target-version = ['py312']

[tool.ruff]
target-version = "py312"
line-length = 88
select = [
    "E", "F", "W",  # flake8 errors, pyflakes, warnings
    "I",            # isort
    "UP",           # pyupgrade
]
ignore = []

[tool.ruff.isort]
known-first-party = ["src"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = "-v --tb=short"
pythonpath = ["."]

[tool.mypy]
python_version = "3.12"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
check_untyped_defs = true
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_no_return = true
warn_unreachable = true
strict_equality = true